requirements.txt
Flask
biopython
pandas
gunicorn